+++ packages/cli/src/greet.ts
