+++ packages/cli/src/help.ts
