+++ packages/core/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@mng-tool/core",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "dependencies": {
+    "@linear/sdk": "^63.2.0",
+    "@octokit/rest": "^22.0.1"
+  },
+  "peerDependencies": {
+    "typescript": "^5"
+  }
+}
