+++ tests/time.test.ts
@@ -1,5 +1,9 @@
 import { describe, expect, test } from "bun:test";
-import { convertLocalDateTimeToUtc, parseLocalDateTimeInput, resolveTimeZone } from "../src/time";
+import {
+  convertLocalDateTimeToUtc,
+  parseLocalDateTimeInput,
+  resolveTimeZone,
+} from "../packages/core/src/time";
 
 describe("parseLocalDateTimeInput", () => {
   test("parses YYYYMMDD0000 values", () => {
